Add a faulty test scenario
This is a test for the quite unlikely situation, that one hosts a
project with a same name on multiple gerrit instances. Let's assume that
project A is hosted on gerrit X and simultaneously project A is hosted
on gerrit Y. When AX already contains a change 1 and a change with the
same number is then created on AY, it will happen that the AX could be
matched instead of AY.

Unfortunataly, this test does not prove it's concept fully, as the
upstream repositories are initiated globally in the tests, for all
connections. For this test to be more valid, we would need to
instantiate each repository for each connection, or at least for each
connection where the project is configured.

Thinking of this, it would benefit by creating only the needed project
for each layout.

Change-Id: Iecab560fa44925abfd15e50555cc3b6dd4a41e93